Implement alternative approach.
This approach uses the Requests library to run the error scraping
performed in `plugins.php`.

The response's body is checked and, if an error is detected,
the plugin is rolled back to the previously installed version.

An email will also be sent to notify the site admin that
an error was encountered and the rollback took place.